Fixed values not being set initially.
Had to use "taskValue()" instead of "taskValue".
Using $inject also didn't work and couldn't find a good guide about best
practices regarding value definition...yet. For now, using a function to
return the actual object. 

This was causing a glitch where, instead of receiving the value defined
in task.value.js, which was " [] ", it was receiving "undefined"
instead.